the cell can see your shooting now

+++ src/Rasa.Game/Structures/AutoFireTimer.cs
@@ -5,11 +5,11 @@
     public class AutoFireTimer
     {
         public Client Client { get; set; }
-        public int RefireTime { get; set; }
+        public long RefireTime { get; set; }
         public long MaxAliveTime { get; set; }
         public long Delay { get; set; }
 
-        public AutoFireTimer(Client client, int refireTime, long delay)
+        public AutoFireTimer(Client client, long refireTime, long delay)
         {
             Client = client;
             RefireTime = refireTime;
